requirements.txt
Flask==0.12.2
Pillow==6.1.0
